+++ Source/wunthshin/Subsystem/ItemSubsystem/ItemSubsystem.h
@@ -5,8 +5,7 @@
 #include "CoreMinimal.h"
 #include "Subsystems/GameInstanceSubsystem.h"
 
-#include "wunthshin/Subsystem/TableQueryEditorSubsystem.h"
-#include "wunthshin/Subsystem/TableQueryGameInstanceSubsystem.h"
+#include "wunthshin/Subsystem/TableQuerySubsystem.h"
 #include "ItemSubsystem.generated.h"
 
 /**
